added stub

+++ YieldView.API/Models/SP500Price.cs
@@ -4,5 +4,5 @@ public class SP500Price
 {
   public int Id { get; set; }
   public DateTime Date { get; set; }
-  public decimal Close { get; set; }
+  public double Close { get; set; }
 }
